Pitching_Model/requirements.txt

psycopg2-binary==2.9.9
numpy==1.26.0
matplotlib==3.6.3
pandas==2.1.1
